Remove slight bias from the RNG
I discovered through keen observation that the RNG had a small bias. Namely, it threw away duplicate results, which means that an adversary who kept track of the generated numbers could be increasingly certain of the next output of the RNG. In fact, after 9007199254740991 samples, the RNG becomes a completely deterministic number generator, a DNG.

This PR improves on the original design of the True Random Number Generator (TRNG), making it a Uniform True Pseudorandom Number Generator (UTPRNG). It uses the Mersenne Twister to generate pseudorandom numbers, and does this in a manner completely compatible with the original API, minus the vulnerability.

You're welcome.